initial wall-jump behavior
Using four ray-casts on each side of the player, to detect walls and distinguish from steps/ledges.
Not 100% happy with the jump trajectory and how it interacts with inputs right now, but fixing that can come later.
Learned a bit about writing macros in Rust, which I used to implement some traits related to the `Side` enum I made for this feature.